Add simple test program for FSGetErrorText()
When run with arguments, prints messages for the given codes.
When run with no arguments, prints messages for known valid error codes
and some invalid codes.

Not run as part of "make check" since it requires a connection to an
xfs server to run, just available for developers to manually run when
needed.

Signed-off-by: Alan Coopersmith <alan.coopersmith@oracle.com>