Avoid reading outside bounds when _FSReply receives an Error response
Upon receiving a response, _FSReply copies the first 8 bytes into *rep
and then looks at them to determine what type of response.   If it's an
error packet, it then converts to an error struct and reads the rest,
but it was copying 16 bytes out of *rep to begin with, due to sloppy
casting.   Since we immediately overwrite the second 8 bytes with the
data coming off the wire, this isn't horrible, but it really freaks out
static analysis and memory debugging tools.

Fixes parfait 1.1 warning:

Error: Buffer overrun
   Read Outside Array Bounds in STD C function: Read outside array bounds in call to llvm.memcpy.p0i8.p0i8.i64. Buffer ((char*)((union fsError*)rep)) of size ??? is read at an offset of 16
      size(((char*)((union fsError*)rep))) is 8, 16 is 16
        at line 751 of src/FSlibInt.c in function '_FSReply'.
        called at line 67 of src/FSSync.c in function 'FSSync' with rep = ((union fsReply*)&rep).

Signed-off-by: Alan Coopersmith <alan.coopersmith@oracle.com>